# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jeremiah Generation Youth API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /membership-requests:
    get:
      operationId: listMembershipRequests
      tags: [membership]
      summary: List membership requests (leader/admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MembershipRequest"
    post:
      operationId: createMembershipRequest
      tags: [membership]
      summary: Submit a membership request (visitor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipRequestInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipRequest"
  /membership-requests/{id}/approve:
    post:
      operationId: approveMembershipRequest
      tags: [membership]
      summary: Approve a membership request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipRequest"
  /membership-requests/{id}/reject:
    post:
      operationId: rejectMembershipRequest
      tags: [membership]
      summary: Reject a membership request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipRequest"

  # Leaders
  /leaders:
    get:
      operationId: listLeaders
      tags: [leaders]
      summary: List leaders and their permissions (super admin only)
      responses:
        "200":
          description: Leaders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderWithPermissions"
    post:
      operationId: addLeader
      tags: [leaders]
      summary: Add a leader (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaderInput"
      responses:
        "201":
          description: Leader added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderWithPermissions"
  /leaders/{profileId}:
    patch:
      operationId: updateLeaderPermissions
      tags: [leaders]
      summary: Update leader permissions (super admin only)
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaderPermissionsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderWithPermissions"
    delete:
      operationId: revokeLeaderAccess
      tags: [leaders]
      summary: Revoke leader access (super admin only)
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /leaders/verify-pin:
    post:
      operationId: verifyLeaderPin
      tags: [leaders]
      summary: Verify leader PIN for dashboard access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinVerification"
      responses:
        "200":
          description: PIN verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinVerificationResult"
        "401":
          description: Invalid PIN
  /leaders/update-pin:
    post:
      operationId: updateLeaderPin
      tags: [leaders]
      summary: Update own PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinUpdate"
      responses:
        "200":
          description: PIN updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # Check-in search
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, full_name, role, created_at]
      properties:
        id:
          type: string
        clerk_id:
          type: ["string", "null"]
        full_name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: [male, female, other, "null"]
        age:
          type: ["integer", "null"]
        heard_from:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, leader, member, visitor]
        created_at:
          type: string
          format: date-time
...
    MembershipRequest:
      type: object
      required: [id, profile_id, reason, status, created_at]
      properties:
        id:
          type: string
        profile_id:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewed_by:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        profile:
          $ref: "#/components/schemas/Profile"
    MembershipRequestInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    LeaderWithPermissions:
      type: object
      required: [profile_id, can_create_events, can_manage_members, can_view_kpis, can_approve_membership]
      properties:
        profile_id:
          type: string
        can_create_events:
          type: boolean
        can_manage_members:
          type: boolean
        can_view_kpis:
          type: boolean
        can_approve_membership:
          type: boolean
        profile:
          $ref: "#/components/schemas/Profile"
    LeaderInput:
      type: object
      required: [profile_id, pin]
      properties:
        profile_id:
          type: string
        pin:
          type: string
        can_create_events:
          type: boolean
          default: true
        can_manage_members:
          type: boolean
          default: false
        can_view_kpis:
          type: boolean
          default: true
        can_approve_membership:
          type: boolean
          default: false
    LeaderPermissionsUpdate:
      type: object
      properties:
        can_create_events:
          type: boolean
        can_manage_members:
          type: boolean
        can_view_kpis:
          type: boolean
        can_approve_membership:
          type: boolean
    PinVerification:
      type: object
      required: [phone, pin]
      properties:
        phone:
          type: string
        pin:
          type: string
    PinVerificationResult:
      type: object
      required: [success, role]
      properties:
        success:
          type: boolean
        role:
          type: string
          enum: [super_admin, leader]
        profile_id:
          type: string
    PinUpdate:
      type: object
      required: [current_pin, new_pin]
      properties:
        current_pin:
          type: string
        new_pin:
          type: string
    SuccessMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
...