# lib/api-zod/src/generated/types/leaderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeremiah Generation Youth API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderInput {
  profile_id: string;
  pin: string;
  can_create_events?: boolean;
  can_manage_members?: boolean;
  can_view_kpis?: boolean;
  can_approve_membership?: boolean;
}