# lib/api-zod/src/generated/types/pinVerification.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeremiah Generation Youth API
 * OpenAPI spec version: 0.1.0
 */

export interface PinVerification {
  phone: string;
  pin: string;
}